+++ services/wallets/src/domain/entities/wallet/index.ts
@@ -1,10 +1,12 @@
+import type { Decimal } from "@prisma/client/runtime/client";
+
 import { Player } from "../player";
 import { Transaction } from "../transaction";
 
 interface WalletProps {
   id: string;
   playerId: string;
-  balance: bigint;
+  balance: Decimal;
   createdAt: Date;
   updatedAt: Date;
   player: Player | null;
@@ -14,7 +16,7 @@ interface WalletProps {
 export class Wallet implements WalletProps {
   id: string;
   playerId: string;
-  balance: bigint;
+  balance: Decimal;
   createdAt: Date;
   updatedAt: Date;
   player: Player | null;
